+++ game4.10/Source/GameMap.h
@@ -20,7 +20,7 @@ namespace game_framework {
 		void OnShowWall();
 		void OnShowPressF();
 		int GetCharacterStatus();
-		int* GetCharacterPosition();
+		float* GetCharacterPosition();
 		void CharacterUseSkill(int, int, int);
 		void OnShow();
 		void SkillOnMove();
@@ -29,15 +29,15 @@ namespace game_framework {
 		void CleanMemory();
 		virtual void LoadBitmap()=0;
 		virtual void OnMove() = 0;
-		virtual int* SetCharacterXY(int, int, const int*) = 0;
+		virtual float* SetCharacterXY(int, int, const int*) = 0;
 		virtual bool SetEnemyXY(int, int, int*) = 0;
-		virtual int GetMapStatus(int, int) = 0;
+		virtual int GetMapStatus(float, float) = 0;
 		
 	protected:
 		CMovingBitmap _background; //背景圖
 		CMovingBitmap _wall; //牆	
 		CAnimation _ani_press_f;  //press f 動畫
-		int _cxy[2] = { 0,0 };	//角色所在Map的座標
+		float _cxy[2] = { 0,0 };	//角色所在Map的座標
 		bool _isPressF;
 		int _character_status;
 		//const int _collision_move[4] = { 24, 53, 20 , 7};   //(x,y,l,w) collision_move	碰撞判定範圍
