+++ game4.10/Source/Map_Level_One.h
@@ -10,9 +10,9 @@ namespace game_framework {
 		~Map_Level_One();
 		void LoadBitmap();
 		void OnMove();
-		int* SetCharacterXY(int, int, const int*);
+		float* SetCharacterXY(int, int, const int*);
 		bool SetEnemyXY(int, int, int*);
-		int GetMapStatus(int, int);
+		int GetMapStatus(float, float);
 	};
 
 }
